app/requirements.txt
sentence-transformers==2.7.0
faiss-cpu==1.7.4
numpy==1.26.4